test(L0): promote shared metarepo_workspace fixture; exercise plan→dev gate on metarepo
